(PC-22881)[PRO] refactor: clean icon full-right & full-left
