+++ pro/src/screens/OfferType/ActionsBar/ActionsBar.tsx
@@ -4,7 +4,7 @@ import ActionsBarSticky from 'components/ActionsBarSticky'
 import { Events } from 'core/FirebaseEvents/constants'
 import { computeOffersUrl } from 'core/Offers/utils'
 import useAnalytics from 'hooks/useAnalytics'
-import { ReactComponent as IcoMiniArrowRight } from 'icons/ico-mini-arrow-right.svg'
+import { ReactComponent as RightIcon } from 'icons/full-right.svg'
 import { ButtonLink, SubmitButton } from 'ui-kit'
 import { ButtonVariant, IconPositionEnum } from 'ui-kit/Button/types'
 
@@ -30,7 +30,7 @@ const ActionsBar = ({
       </ActionsBarSticky.Left>
       <ActionsBarSticky.Right>
         <SubmitButton
-          Icon={IcoMiniArrowRight}
+          Icon={RightIcon}
           iconPosition={IconPositionEnum.RIGHT}
           disabled={disableNextButton}
         >
