+++ pro/src/ui-kit/Button/ButtonLink.tsx
@@ -2,7 +2,8 @@ import cn from 'classnames'
 import React, { MouseEventHandler, useId } from 'react'
 import { Link } from 'react-router-dom'
 
-import { ReactComponent as IcoArrowRight } from 'icons/ico-mini-arrow-right.svg'
+import fullRight from 'icons/full-right.svg'
+import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
 import Tooltip from 'ui-kit/Tooltip'
 
 import styles from './Button.module.scss'
@@ -71,7 +72,9 @@ const ButtonLink = ({
       {
         /* istanbul ignore next: graphic variation */ variant ===
           ButtonVariant.BOX && (
-          <IcoArrowRight
+          <SvgIcon
+            src={fullRight}
+            alt=""
             className={cn(styles['button-icon'], styles['button-icon-arrow'])}
           />
         )
